+++ tests/testsuite_mb_serial/test_mb_s_iterator_filteredsorted.c
@@ -91,7 +91,7 @@ void test_mb_s_iter_create_filteredsorted(void) {
     iterator = (MBIt_Iterator *)MBI_getIteratorRef(itr_fs);
     CU_ASSERT_PTR_NOT_NULL_FATAL(board);
     CU_ASSERT_PTR_NOT_NULL_FATAL(iterator);
-    CU_ASSERT_EQUAL((int)board->data->elem_size, iterator->msgsize);
+    CU_ASSERT_EQUAL(board->data->elem_size, iterator->msgsize);
     CU_ASSERT_EQUAL(iterator->iterating, 0);
     CU_ASSERT_PTR_NULL(iterator->cursor);
     CU_ASSERT_EQUAL(iterator->mb, mb_fs);
